+++ MUS4_FW/libraries/mus4_wifi/src/WifiManager.h
@@ -47,6 +47,7 @@ String getActiveWifiApSsid();
 
 void setupWifiConsole();
 void updateWifiSta();
+void updateWifiStaHistoryRetry();
 void updateWifiConsole();
 
 void setupWifiWebConsole();