+++ MUS4_FW/libraries/mus4_wifi/src/mus4_wifi.h
@@ -14,4 +14,5 @@
 #include "WifiManager.h"
 #include "WifiOta.h"
 #include "WifiStaConfig.h"
+#include "WifiStaHistory.h"
 #include "WifiIdentity.h"